requirements.txt
flask==2.3.3
flask-socketio==5.3.6
flask-cors==4.0.0
flask-login==0.6.3
flask-bcrypt==1.0.1
python-socketio==5.8.0
python-engineio==4.5.1
openai==1.99.9
simple-websocket==1.0.0
bcrypt==4.0.1
pymongo==4.6.1
python-dotenv==1.0.0
